requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
systemd-python==235
mypy==1.9.0
PyYAML==6.0.1